refactor(stats,dependency-leases): review fixes (#677)
- Move bar() ownership to stats-index.ts: the facade re-exports
  stats-index, so stats-index importing bar from ./stats was a
  facade<->extracted-module cycle; bar was also never part of the
  public surface, so the facade's export of it has been dropped.
- dependency-leases facade: import type for DependencyLeasePaths /
  AcquireDependencyInput (type-only symbols were imported value-form).
